fix: extract stat collection functions
