+++ src/service/search/utils.rs
@@ -15,12 +15,13 @@
 
 use std::{future::Future, pin::Pin, sync::Arc};
 
-use config::meta::search::ScanStats;
+use config::meta::{cluster::NodeInfo, search::ScanStats};
 use datafusion::physical_plan::{ExecutionPlan, ExecutionPlanVisitor};
 use sqlparser::ast::{BinaryOperator, Expr};
 use tokio::sync::Mutex;
 
 use super::{DATAFUSION_RUNTIME, datafusion::distributed_plan::remote_scan::RemoteScanExec};
+use crate::service::grpc::make_grpc_search_client;
 
 type Cleanup = Pin<Box<dyn Future<Output = ()> + Send>>;
 
@@ -166,4 +167,36 @@ pub fn is_value(e: &Expr) -> bool {
 
 pub fn is_field(e: &Expr) -> bool {
     matches!(e, Expr::Identifier(_) | Expr::CompoundIdentifier(_))
+}
+
+pub async fn collect_scan_stats(
+    nodes: &[Arc<dyn NodeInfo>],
+    trace_id: &str,
+    is_leader: bool,
+) -> ScanStats {
+    let mut scan_stats = ScanStats::default();
+    for node in nodes {
+        let mut scan_stats_request = tonic::Request::new(proto::cluster_rpc::GetScanStatsRequest {
+            trace_id: trace_id.to_string(),
+            is_leader,
+        });
+        let mut client =
+            match make_grpc_search_client(trace_id, &mut scan_stats_request, node).await {
+                Ok(c) => c,
+                Err(e) => {
+                    log::error!("error in creating get scan stats client :{e}, skipping");
+                    continue;
+                }
+            };
+        let stats = match client.get_scan_stats(scan_stats_request).await {
+            Ok(v) => v,
+            Err(e) => {
+                log::error!("error in getting scan stats : {e}, skipping");
+                continue;
+            }
+        };
+        let stats = stats.into_inner().stats.unwrap_or_default();
+        scan_stats.add(&(&stats).into());
+    }
+    scan_stats
 }
